refactor rendering out of InstrumentActor #21582
